# KDTree neighbour search in `filter_min_distance`