Modifies `filter_min_distance` to use a KDTree neighbour search to resolve the feature order dependence highlighted in #230 . This also allows `filter_min_distance` to remove child features of features that have already been removed, unlike the previous implementation (although this is a rare situation in testing)

As a benefit, this has also provided a nice speedup to the function. A comparison for running over a list of 629 features:

Old version: 8.93 s ± 446 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

New version: 533 ms ± 52.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

PR created as a draft awaiting #244 

* [x] Have you followed our guidelines in CONTRIBUTING.md? 
* [x] Have you self-reviewed your code and corrected any misspellings? 
* [x] Have you written documentation that is easy to understand?
* [x] Have you written descriptive commit messages? 
* [x] Have you added NumPy docstrings for newly added functions? 
* [x] Have you formatted your code using black? 
* [ ] If you have introduced a new functionality, have you added adequate unit tests?
* [x] Have all tests passed in your local clone? 
* [ ] If you have introduced a new functionality, have you added an example notebook?
* [x] Have you kept your pull request small and limited so that it is easy to review? 
* [x] Have the newest changes from this branch been merged?