@snilsn 2023-02-17
The speedup sounds great! Given my relation to #230 , I would like to do a review here.