@freemansw1 2023-05-22
> @snilsn @freemansw1 this is now ready for review. It is missing PBC however, Sean did you have a KDTree implementation that worked for tracking?

As I mentioned in the meeting this morning, I had to use a BTree for the PBC implementation of tracking. I used the BTree implementation in scikit-learn: https://scikit-learn.org/stable/modules/generated/sklearn.neighbors.BallTree.html . We now have sklearn as a requirement, so this is feasible.

I would be inclined to only use the BTree for the PBC case, but I could see the advantage the other way as well. To call the BTree for the PBC case, you first need to create a distance function (see https://github.com/tobac-project/tobac/blob/e74eaa64857ed36bc949c334466a84ef6f8eb2dd/tobac/tracking.py#LL617C1-L617C1 and https://github.com/tobac-project/tobac/blob/e74eaa64857ed36bc949c334466a84ef6f8eb2dd/tobac/tracking.py#LL335C1-L336C1 for how I did this in tracking). To call the BTree with a python distance function (which I don't think is the long-term endgame, it would be better to create a custom `DistanceMetric` for improved PBC performance instead of having the python calling overhead):

`BallTree(coords_mapped, metric='pyfunc', func=self.dist_func)`