@w-k-jones 2023-05-22
Thanks Sean, that's really useful! I'll look into implementing it