@w-k-jones 2023-05-25
The failed tests were due to the lack of PBC, so I implemented the ball tree option as per @freemansw1 's advice. Only difficulty was that Balltree returns a numpy array of numpy arrays, rather than a list of lists, so I made the slightly hacky conversion in line 1435 https://github.com/tobac-project/tobac/blob/5401826e08e277fd07d38adf89426ba852f9a400/tobac/feature_detection.py#L1435

Now ready for review